Revert "Fixed output type"
This reverts commit 458a02425162f83fb7a74f0455f7ab8b36819a43.

+++ src/main/java/org/spout/api/util/hashing/TByteTripleHashed.java
@@ -45,8 +45,8 @@ public class TByteTripleHashed {
 	 * @param key to get from
 	 * @return the first 8-bit integer value in the key
 	 */
-	public static final byte key1(int key) {
-		return (byte) (key >> 16 & 0xFF);
+	public static final int key1(int key) {
+		return key >> 16 & 0xFF;
 	}
 
 	/**
@@ -55,8 +55,8 @@ public class TByteTripleHashed {
 	 * @param key to get from
 	 * @return the second 8-bit integer value in the key
 	 */
-	public static final byte key2(int key) {
-		return (byte) (key & 0xFF);
+	public static final int key2(int key) {
+		return key & 0xFF;
 	}
 
 	/**
@@ -65,7 +65,7 @@ public class TByteTripleHashed {
 	 * @param key to get from
 	 * @return the third 8-bit integer value in the key
 	 */
-	public static final byte key3(int key) {
-		return (byte) (key >> 8 & 0xFF);
+	public static final int key3(int key) {
+		return key >> 8 & 0xFF;
 	}
 }
